Handle fragment to make edit pages bookmarkable. Refactorize.
The filtering state of the quote list should be bookmarkable too but
this requires more effort than it is worth because jsGrid doesn't
update the visual representation of the filter when you set it.

The refactoring puts the code for each page in a separate module and
groups the actions in two controllers (api and quotes).